+++ FOX/classes/__init__.py
@@ -1,9 +1,9 @@
 from .multi_mol import MultiMolecule
-from .monte_carlo import MonteCarlo
+from .monte_carlo import MonteCarlo, ARMC
 from .molecule_utils import Molecule
 
 __all__ = [
     'MultiMolecule',
-    'MonteCarlo',
+    'MonteCarlo', 'ARMC',
     'Molecule'
 ]