+++ CHANGELOG.rst
@@ -5,6 +5,23 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+0.2.1
+*****
+
+Added
+-----
+
+* Introduced new MutliMolecule methods for slicing MD trajectories
+* Added the MonteCarlo API to the documentation
+* WiP: Split the MonteCarlo class into 2 classes: MonteCarlo & ARMC (subclass)
+
+Changed
+-------
+
+* Minor update to copy/deepcopy-related methods
+* Improved the get_rdf() function
+
+
 0.2
 ***